+++ src/components/container/Header/header.module.css
@@ -1,6 +1,6 @@
 .header {
-    margin-top: auto;
     color:darkmagenta;
     font-family: 'Amatic SC', cursive;
+    font-weight: bolder;
     font-size: 80px;
 }
