--- src/components/container/add/add.js
@@ -1,32 +1,0 @@
-// import React from 'react';
-// import styles from './add.module.css';
-// import {useRef} from 'react'
-// import List from '../list/list';
-
-
-// function Add(){
-   
-//     const addItem = (e) =>{
-//        let copy = [...List.itemsList]
-//        copy = [...copy, e.currentTarget.value]
-//        List.setItemsList(copy);
-//     }
-//     return (
-//         <div>
-//             <div>
-//             <input 
-//             placeholder="Type an item..."
-//             className={styles.input}
-//             type="text"
-//             ref={value}
-//             /></div>
-//             <div>
-//             <button 
-//                 onClick={addItem}
-//                 className={styles.button}> Add product </button>
-//             </div>
-//         </div>
-//     )
-// }
-
-// export default Add;
