+++ src/components/container/container.module.css
@@ -1,7 +1,5 @@
 .Container {
-    text-align: center;
-    margin-left: auto;
-    margin-right: auto;
+    margin: auto;
     margin-top: 20px;
     margin-bottom: 20px;
     min-height: 900px;
