--- src/components/container/list/add.module.css
@@ -1,30 +1,0 @@
-.Add {
-    margin-top: 10px;
-    margin-bottom: 20px;
-    display: flex;
-}
-.input {
-    height: 30px;
-    min-width: 450px;
-    border: solid 2px slategray;
-    border-radius: 5px;
-    font-size: 20px;
-    padding-left: 10px;
-}
-.button {
-
-    height: 50px;
-    width: 200px;
-    border: solid 3px limegreen;
-    border-radius: 60px;
-    font-size: 20px;
-    font-weight:bold;
-    background-color: limegreen;
-    color: white;
-    margin-top: 10px;
-}
-.button:hover{
-    background-color:lime;
-    border: solid 3px lime;
-    color:white;
-}
