+++ src/components/container/list/list.js
@@ -1,32 +1,17 @@
 import React, { useEffect } from 'react';
 import styles from './list.module.css';
 import { useState } from 'react';
+import Input from './input/input';
 
 
 const List = () => {
     const [itemsList, setItemsList] = useState([]);
-    const [userInput, setUserInput] = useState();
 
     useEffect(() => {
         let listEl = localStorage.getItem("list");
         let list = JSON.parse(listEl ?? []);
         setItemsList(list);
     },[]);
-
-    const addItem = (e) => {
-        let copy = [...itemsList]
-        copy = [...copy, {
-            "id": copy.length - 1,
-            "task": userInput,
-            "complete": false
-        }]
-        setItemsList(copy);
-        localStorage.setItem("list", JSON.stringify(copy))
-    }
-
-    const handleChange = (e) => {
-        setUserInput(e.currentTarget.value)
-    }
 
     const deleteItem = (e) => {
         let copy = [...itemsList]
@@ -35,7 +20,6 @@ const List = () => {
         setItemsList(copy);
         localStorage.setItem("list", JSON.stringify(copy))
     }
-
 
     const handleClick = (e) => {
         e.preventDefault()
@@ -53,29 +37,18 @@ const List = () => {
 
     return (
         <div>
-            <div>
-                <div>
-                    <input
-                        value={userInput} onChange={handleChange}
-                        placeholder="Type an item..."
-                        className={styles.input}
-                        type="text"
-                    /></div>
-                <div>
-                    <button
-                        onClick={(e) => addItem(e)}
-                        className={styles.button}> Add product </button>
-
-                </div>
-            </div>
+                <Input 
+                    itemsList={itemsList}
+                    setItemsList={setItemsList}
+                />
 
             <ul className={styles.list}>
                 {itemsList.map((el, idx) => (
                     <li key={idx} className={el.complete ? styles.strike : ""}>
-                        <button id={el.id} onClick={handleClick}>
-                            x </button>
-                        {el.task}
-                        <button id={idx} onClick={deleteItem}>Delete</button></li>
+                        <button id={el.id} onClick={handleClick} className={styles.complete}>
+                        ✓ </button>
+                        <button id={idx} onClick={deleteItem} className={styles.delete}>✘</button>
+                        {el.task}</li>
                 ))}
 
             </ul>
