+++ src/variables.module.css
@@ -1,0 +1,9 @@
+.App{
+    color: white;
+    font-size: 20px;
+    text-align: center;
+    padding-left: 20px;
+    padding-top: 20px;
+    font-weight: bolder;
+    font-family: 'Kalam', cursive;
+}
